Add Fate Engine, server integration, and Fate Control WinForms client

+++ Nri.FateControlClient/Models/FateLayerRow.cs
@@ -6,6 +6,6 @@ public sealed class FateLayerRow
     public string DisplayName { get; set; } = string.Empty;
     public bool Enabled { get; set; }
     public int FlatModifier { get; set; }
-    public int Intensity { get; set; }
+    public double Intensity { get; set; }
     public string Mode { get; set; } = "flat";
 }
